requirements.txt
psycopg2~=2.9.10
python-dotenv~=1.1.0
numpy~=1.26.4
pandas~=2.2.3
plotly~=6.0.1
joblib~=1.4.2
tensorflow~=2.19.0
Flask~=3.0.3
scikit-learn~=1.6.1
openai~=1.75.0
pyproj~=3.7.1